회원가 정보가 일치할때만 글 삭제되게끔 설정

+++ pages/api/post/delete.js
@@ -1,13 +1,27 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(req, res) {
   if (req.method == "POST") {
+    let session = await getServerSession(req, res, authOptions);
     console.log(req.body);
-    let db = (await connectDB).db("forum");
-    let result = db
+
+    const db = (await connectDB).db("forum");
+    let idMatch = await db
       .collection("post")
-      .deleteOne({ _id: new ObjectId(req.body) });
-    res.redirect(200).JSON('삭제하였습니다.');
+      .findOne({ _id: new ObjectId(req.body) });
+
+    console.log(idMatch);
+    if (idMatch.author == session.user.email) {
+      let result = await db
+        .collection("post")
+        .deleteOne({ _id: new ObjectId(req.body) });
+
+        return res.status(200).json({ message: '삭제완료' })
+    } else {
+      return res.status(500).json({ message: '유저와 작성자 불일치' });
+    }
   }
 }
